AAAAAHHHHHHHHHHH
IT WORKS!!! Able to switch panes and click on buttons to view data for each weapon class. Minor bug in that elements on the hello-view fxml cannot be used. Attempted to modify the text in the input field and it was disabled after clicking anywhere that would be considered part of the WeaponTable fxml